FantasyTeamPage Updates (see details)
Redux persistence added to allow email to be stored. Email storage required as this attribute is needed to fetch a specific users (via email) list of player IDs stored in the selectedPlayersID string attribute.

Fetch/Reset fantasy team actions added. These work with the bankend to fetch and display a users team if they already have one. An option is also added to reset their team, this will refresh the page and dispaly the new blank team.